refactor: remove dead helpers and share the replay test config
detect_encoding, format_sys_table_memory, MetricManager.trim_datetimes_to_window,
GraphTab.metric_groups, and Dolphie.statements_summary_metrics had no callers.
detect_encoding was the only import of charset_normalizer, which pyproject never
declared.

replay_config and REPLAYS move to the test harness so the playback and snapshot
tests build the same Config. action_replay_forward reports the end of the replay
from seek_relative like action_replay_back does, and the held-key test posts
repeats until the cursor reaches the last frame instead of a fixed count.
